requirements.venv.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
flake8>=3.9.2,<3.10
